# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trackvance Core",
    "description": "Monolito modular local. SUCCESS es estado técnico; decision conserva el resultado de negocio. Configuraciones y evidence son snapshots versionados.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/findings/{finding_id}/exceptions": {
      "post": {
        "summary": "Finding Exception",
        "operationId": "finding_exception_api_v1_findings__finding_id__exceptions_post",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Finding Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "exceptions:write"
      }
    },
...
    "/api/v1/exceptions": {
      "get": {
        "tags": [
          "Excepciones"
        ],
        "summary": "Exceptions",
        "operationId": "exceptions_api_v1_exceptions_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "module",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Module"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "assigned_user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned User Id"
            }
          },
          {
            "name": "overdue",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Overdue"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "exceptions:read"
      }
    },
    "/api/v1/exceptions/assignees": {
      "get": {
        "tags": [
          "Excepciones"
        ],
        "summary": "Assignees",
        "operationId": "assignees_api_v1_exceptions_assignees_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "exceptions:read"
      }
    },
    "/api/v1/exceptions/{case_id}": {
      "get": {
        "tags": [
          "Excepciones"
        ],
        "summary": "Detail",
        "operationId": "detail_api_v1_exceptions__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "exceptions:read"
      },
      "patch": {
        "tags": [
          "Excepciones"
        ],
        "summary": "Update Case",
        "operationId": "update_case_api_v1_exceptions__case_id__patch",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CasePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "exceptions:write"
      }
    },
    "/api/v1/exceptions/{case_id}/validate": {
      "post": {
        "tags": [
          "Excepciones"
        ],
        "summary": "Validate",
        "operationId": "validate_api_v1_exceptions__case_id__validate_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "exceptions:write"
      }
    },
    "/api/v1/exceptions/{case_id}/comments": {
      "post": {
        "tags": [
          "Excepciones"
        ],
        "summary": "Comment",
        "operationId": "comment_api_v1_exceptions__case_id__comments_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sesión ausente o vencida"
          },
          "403": {
            "description": "Permiso u origen no autorizado / CSRF inválido"
          },
          "404": {
            "description": "Recurso no disponible en la organización"
          },
          "409": {
            "description": "Conflicto de versión, precondición o integridad"
          }
        },
        "security": [
          {
            "LocalSession": []
          }
        ],
        "x-required-permission": "exceptions:write"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommentBody": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "comment": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "comment"
        ],
        "title": "CommentBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationBody": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          },
          "validation_run_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Run Id"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version"
        ],
        "title": "ValidationBody"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "LocalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "trackvance_session",
        "description": "Sesión local HttpOnly. Las mutaciones requieren además X-CSRF-Token."
      }
    }
  }
}